requirements.txt
django 
djangorestframework 
psycopg2-binary 
celery 
redis 
pandas 
numpy 
scikit-learn 
tensorflow